Conflicts:
	charms/oscied-orchestra/local_config.pkl
	charms/oscied-orchestra/revision
	charms/oscied_lib/Orchestra.py
	charms/oscied_lib/OrchestraConfig.py
	charms/oscied_lib/PublishTask.py
	charms/oscied_lib/TransformTask.py
	scripts/menu.sh

+++ charms/oscied_lib/OrchestraHooks.py
@@ -28,9 +28,8 @@ import os, re, shutil, time
 from codecs import open
 from kitchen.text.converters import to_bytes
 from configobj import ConfigObj
-from CharmHooks import DEFAULT_OS_ENV
-from CharmHooks_Storage import CharmHooks_Storage
-from OrchestraConfig import OrchestraConfig
+from oscied_config import OrchestraLocalConfig
+from oscied_hook_base import CharmHooks_Storage, DEFAULT_OS_ENV
 from pyutils.py_filesystem import first_that_exist, try_makedirs
 from pyutils.py_subprocess import rsync, screen_launch, screen_list, screen_kill
 
@@ -39,11 +38,11 @@ class OrchestraHooks(CharmHooks_Storage):
 
     PACKAGES = tuple(set(CharmHooks_Storage.PACKAGES +
                      (u'ffmpeg', u'ntp', u'x264', u'mongodb', u'rabbitmq-server')))
-    JUJU_PACKAGES = (u'juju', u'juju-jitsu')
+    JUJU_PACKAGES = (u'juju-core',)
 
     def __init__(self, metadata, default_config, local_config_filename, default_os_env):
         super(OrchestraHooks, self).__init__(metadata, default_config, default_os_env)
-        self.local_config = OrchestraConfig.read(local_config_filename, store_filename=True)
+        self.local_config = OrchestraLocalConfig.read(local_config_filename, store_filename=True)
         self.debug(u'My __dict__ is {0}'.format(self.__dict__))
 
     # ------------------------------------------------------------------------------------------------------------------
@@ -56,12 +55,12 @@ class OrchestraHooks(CharmHooks_Storage):
         return u'mongodb://admin:{0}@localhost:27017/orchestra'.format(self.config.mongo_admin_password)
 
     @property
-    def mongo_nodes_connection(self):
-        return u'mongodb://nodes:{0}@{1}:27017/celery'.format(self.config.mongo_nodes_password, self.private_address)
+    def mongo_node_connection(self):
+        return u'mongodb://node:{0}@{1}:27017/celery'.format(self.config.mongo_node_password, self.private_address)
 
     @property
     def rabbit_connection(self):
-        return u'amqp://nodes:{0}@{1}:5672/celery'.format(self.config.rabbit_password, self.private_address)
+        return u'amqp://node:{0}@{1}:5672/celery'.format(self.config.rabbit_password, self.private_address)
 
     @property
     def rabbit_users(self):
@@ -79,12 +78,12 @@ class OrchestraHooks(CharmHooks_Storage):
         self.info(u'Configure RabbitMQ Message Broker')
         self.cmd(u'rabbitmqctl delete_user guest', fail=False)
         self.cmd(u'rabbitmqctl delete_vhost /', fail=False)
-        self.cmd(u'rabbitmqctl add_user nodes "{0}"'.format(self.config.rabbit_password), fail=False)
+        self.cmd(u'rabbitmqctl add_user node "{0}"'.format(self.config.rabbit_password), fail=False)
         self.cmd(u'rabbitmqctl add_vhost celery', fail=False)
-        self.cmd(u'rabbitmqctl set_permissions -p celery nodes ".*" ".*" ".*"', fail=False)
+        self.cmd(u'rabbitmqctl set_permissions -p celery node ".*" ".*" ".*"', fail=False)
         users, vhosts = self.rabbit_users, self.rabbit_vhosts
         self.debug(u'RabbitMQ users: {0} vhosts: {1}'.format(users, vhosts))
-        if u'guest' in users or u'nodes' not in users or u'/' in vhosts or u'celery' not in vhosts:
+        if u'guest' in users or u'node' not in users or u'/' in vhosts or u'celery' not in vhosts:
             raise RuntimeError(to_bytes(u'Unable to configure RabbitMQ'))
 
     # ------------------------------------------------------------------------------------------------------------------
@@ -93,7 +92,7 @@ class OrchestraHooks(CharmHooks_Storage):
         self.hook_uninstall()
         self.info(u'Upgrade system and install prerequisites')
         self.cmd(u'apt-add-repository -y ppa:jon-severinsson/ffmpeg')
-        self.cmd(u'apt-add-repository -y ppa:juju/pkgs')
+        self.cmd(u'apt-add-repository -y ppa:juju/stable')
         self.cmd(u'apt-get -y update', fail=False)
         self.cmd(u'apt-get -y upgrade')
         self.cmd(u'apt-get -y install {0}'.format(u' '.join(OrchestraHooks.PACKAGES)))
@@ -119,19 +118,23 @@ class OrchestraHooks(CharmHooks_Storage):
         self.open_port(5672,  u'TCP')  # RabbitMQ service
 
     def hook_config_changed(self):
+        self.info(u'Start MongoDB and RabbitMQ daemons')
+        self.cmd(u'service mongodb start',         fail=False)
+        self.cmd(u'service rabbitmq-server start', fail=False)
+
         self.info(u'Configure Secure Shell')
         rsync(self.local_config.ssh_template_path, self.local_config.ssh_config_path, recursive=True, log=self.debug)
 
         self.info(u'Configure JuJu Service Orchestrator')
-        if not os.path.exists(self.local_config.juju_config_file):
-            try_makedirs(os.path.dirname(self.local_config.juju_config_file))
-            shutil.copy(self.local_config.juju_template_file, self.local_config.juju_config_file)
+        juju_config_path = os.path.dirname(self.local_config.juju_config_file)
+        try_makedirs(juju_config_path)
+        rsync(self.local_config.juju_template_path, juju_config_path, recursive=True, log=self.debug)
 
         self.info(u'Configure MongoDB Scalable NoSQL DB')
         with open(u'f.js', u'w', u'utf-8') as mongo_f:
             mongo_f.write(u"db.addUser('admin', '{0}', false);".format(self.config.mongo_admin_password))
         with open(u'g.js', u'w', u'utf-8') as mongo_g:
-            mongo_g.write(u"db.addUser('nodes', '{0}', false);".format(self.config.mongo_nodes_password))
+            mongo_g.write(u"db.addUser('node', '{0}', false);".format(self.config.mongo_node_password))
         self.cmd(u'mongo f.js')
         self.cmd(u'mongo orchestra f.js')
         self.cmd(u'mongo celery g.js')
@@ -151,13 +154,13 @@ class OrchestraHooks(CharmHooks_Storage):
         self.local_config.api_url = self.api_url(local=False)
         self.local_config.root_secret = self.config.root_secret
         self.local_config.mongo_admin_connection = self.mongo_admin_connection
-        self.local_config.mongo_nodes_connection = self.mongo_nodes_connection
+        self.local_config.mongo_node_connection = self.mongo_node_connection
         self.local_config.rabbit_connection = self.rabbit_connection
         infos = {
             u'rabbit': unicode(self.rabbit_connection),
             u'port': unicode(27017),
-            u'username': u'nodes',
-            u'password': unicode(self.config.mongo_nodes_password),
+            u'username': u'node',
+            u'password': unicode(self.config.mongo_node_password),
         }
         self.template2config(self.local_config.celery_template_file, self.local_config.celery_config_file, infos)
         self.local_config.email_server = self.config.email_server
@@ -175,11 +178,12 @@ b
         #self.cmd('juju destroy-environment')
         #self.cmd('... --purge apt-cacher-ng charm-tools juju libzookeeper-java lxc zookeeper')
         self.storage_unregister()
-        self.cmd(u'apt-get -y remove --purge {0}'.format(u' '.join(OrchestraHooks.PACKAGES)))
-        self.cmd(u'apt-get -y autoremove')
-        #shutil.rmtree('$HOME/.juju $HOME/.ssh/id_rsa*
-        shutil.rmtree(u'/etc/rabbitmq/',     ignore_errors=True)
-        shutil.rmtree(u'/var/log/rabbitmq/', ignore_errors=True)
+        if self.config.cleanup:
+            self.cmd(u'apt-get -y remove --purge {0}'.format(u' '.join(OrchestraHooks.PACKAGES)))
+            self.cmd(u'apt-get -y autoremove')
+            #shutil.rmtree('$HOME/.juju $HOME/.ssh/id_rsa*
+            shutil.rmtree(u'/etc/rabbitmq/',     ignore_errors=True)
+            shutil.rmtree(u'/var/log/rabbitmq/', ignore_errors=True)
         self.local_config.reset()
 
     def hook_start(self):
@@ -221,7 +225,7 @@ b
         # FIXME something to do (register unit ?)
 
     def hook_publisher_relation_joined(self):
-        self.relation_set(mongo_connection=self.mongo_nodes_connection,
+        self.relation_set(mongo_connection=self.mongo_node_connection,
                           rabbit_connection=self.rabbit_connection)
 
     def hook_publisher_relation_changed(self):
@@ -234,7 +238,7 @@ b
         # FIXME something to do (register unit ?)
 
     def hook_transform_relation_joined(self):
-        self.relation_set(mongo_connection=self.mongo_nodes_connection,
+        self.relation_set(mongo_connection=self.mongo_node_connection,
                           rabbit_connection=self.rabbit_connection)
 
     def hook_transform_relation_changed(self):
